pyproject.toml
[project]
name = "mergetune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "inquirer>=3.4.1",
    "metorial>=1.0.11",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
]
